work on timer (2 times slower for an obscure reason :|)

+++ common/src/main/java/org/exodusstudio/stellaris/common/components/TimerComponents.java
@@ -5,14 +5,14 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 
-public record TimerComponents(int timeLeft, int maxTime) {
+public record TimerComponents(double timeLeft, int maxTime) {
     public TimerComponents(int maxTime) {
         this(maxTime, maxTime);
     }
 
     public static Codec<TimerComponents> CODEC = RecordCodecBuilder.create(instance ->
             instance.group(
-                    Codec.INT.fieldOf("timeLeft").forGetter(TimerComponents::timeLeft),
+                    Codec.DOUBLE.fieldOf("timeLeft").forGetter(TimerComponents::timeLeft),
                     Codec.INT.fieldOf("maxTime").forGetter(TimerComponents::maxTime)
             ).apply(instance, TimerComponents::new)
     );
@@ -22,14 +22,14 @@ public record TimerComponents(int timeLeft, int maxTime) {
     static {
         STREAM_CODEC = StreamCodec.of(
                 (buf, timerComponents) -> {
-                    buf.writeInt(timerComponents.timeLeft);
+                    buf.writeDouble(timerComponents.timeLeft);
                     buf.writeInt(timerComponents.maxTime);
                 },
-                buf -> new TimerComponents(buf.readInt(), buf.readInt())
+                buf -> new TimerComponents(buf.readDouble(), buf.readInt())
         );
     }
 
-    public TimerComponents tick(int delta) {
+    public TimerComponents tick(double delta) {
         if (timeLeft - delta > 0) {
             return new TimerComponents(timeLeft - delta, maxTime);
         }
